+++ doc/version-history.rst
@@ -14,6 +14,9 @@ v0.4.0
 ======
 * Swap ts-salobj base case.
 * Fix import path for pymodbus.
+* Add mock_server and simulation mode.
+* Switch to async modbus client.
+* Add sshtunnel functionality
 
 v0.3.0
 ======